Powershell version with new architecture
This new architecture better leverages the AWS CLI and can tell when you are trying to set a session credentials for your mfa'ed user or if you are trying to assume a role. It then also stores the values in the cache like the CLI does and will let you hop back and forth between different roles and users with ease!

The flow is:
Use the passed in profile name to determine if you are trying to use user creds or a role.
This is done by calling aws configure get, checking if the source_profile and role_arn
If you are assuming a role, set the credentials to the source profile temporarily, then call assume-role and set the new credentials.
If you want a user session, set the credentials from the profile name passed in as a param.