Further optimizations and improvements
Changed the `export [environment variable]` to a more correct `unset [environment variable]`.
Changed the `exit 1` to `false` as the former would close the user's shell.
Changed the logic in `Get-ProfileName()` so that when no parameter is passed, `default` is set as the profile name.
Changed the inline python in `Parse-Credentials()` to using the `jp` commands.
Removed the time calculations from `Get-SavedSession()` to speed up the script. This was substituted with a check of both aws's sts calls to verify the credentials have not expired. The script tells the user when each credential is going to expire.
The region is hard-coded to `us-east-1` when making the call to `aws sts get-session-token`
Removed the cache'ing of the role credentials, as it was unnecessary.