Fixes #91: refresh role credentials sooner
The check to refresh role credentials needs to take into account that
they may expire within 60s so that we don't have a period of time where
the role credentials have expired.

This also ensure that if we never sleep for a negative time as this
leads to autoawsume exiting with a failure.